Add TodoAPP_Authentication with JWT authentication requests

+++ TodoAPP_Authentication/routers/todos.py
@@ -5,7 +5,7 @@ from typing import Annotated
 from fastapi import Depends
 from pydantic import BaseModel, Field
 from models import Todo
-
+from routers.auth import get_current_user
 
 router = APIRouter()
 
@@ -19,7 +19,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
-
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -28,29 +28,37 @@ class TodoRequest(BaseModel):
     completed: bool
 
 
-@router.get("/")
-async def root(db: db_dependency):
-    return db.query(Todo).all()
+@router.get("/",status_code=status.HTTP_200_OK)
+async def root(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+    return db.query(Todo).filter(Todo.owner_id == user.get("id")).all()
 
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def get_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
-    if todo is None:
-        raise HTTPException(status_code=404, detail="Todo not found")
-    return todo
+async def get_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+   if user is None:
+       raise HTTPException(status_code=401, detail="Authentication failed")
+   todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get("id")).first()
+   if todo is None:
+       raise HTTPException(status_code=404, detail="Todo not found")
+   return todo
 
 
 @router.post("/todos", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todo(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+    todo_model = Todo(**todo_request.model_dump(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todos/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     todo.title = todo_request.title
@@ -62,8 +70,10 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
 
 
 @router.delete("/todos/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     db.delete(todo)
